Move dlopenflags handling into new library loader
The C library is no longer dynamically loaded directly when `_lupa` is
imported, so we need to move the special handling of dlopenflags into
the `_import_newest_lib()` function in order to ensure they are set when
the lupa C extension is `dlopen`'d.
We also expose a context manager `eager_global_linking()` for users with
special needs to get the same behaviour.